+++ src/components/ForgotPassword/index.js
@@ -1,0 +1,5 @@
+import Step1 from "./Step1";
+
+export {
+    Step1
+}
